Fix for `removeMarkers` doesn't remove all markers #360
This particular problem comes from the pattern
markerClusterer.removeMarkers( markerClusterer.getMarkers() ) as the
underlying mechanism to remove markers performs a splice on the same array
that the remove markers method is working against. This creates a local
copy of the passed in markers array so the above usage pattern works.