Change property name will make markerwithlabel work with markerclusterer
Changing the name to the markerwithlabel this.label property to any other name, like this.markerLabel, will avoid a crash occurs when markerclustererplus invokes setMap. Somehow this.label type changes (property is overwritten) and a "setMap method doesn't exist" error message is displayed. So if in line 526 we assign the MarkerLabel_ to another property name, this error won't happen and the markerclusterer objects will display beautiful.