Create options imageList and commonStyleString
option `imageList` will be useful to declare the list of images path to be used locally as clusters images without thinking in styles. In some cases is necessary to rename the image names in the deploy process to invalidate cache and the images got a random name. 

option `commonStyleString` is used to append some extra common style to each cluster image, like opacity or other style.