refactor(markerclustererplus): replace parseInt with Math.floor
In all this cases parseInt was used on already numbers.
All math operations (/,*,-) cast input type to a number
so it's useless to cast it again with parseInt.
This was OK in JS, but become sensible in TS because of Type System.

parseInt also has side effect, while it parse number-like strings it
omits decimal part, so equal replace will be `Math.floor` instead of
omitting entire parseInt expression.